+++ lib/src/v1/main.dart
@@ -3,7 +3,9 @@ import 'package:meta/meta.dart' show required;
 import '../enums.dart';
 import '../http_helper.dart';
 import '../models/category.dart';
+import '../models/mod.dart';
 import '../models/search.dart';
+import '../models/user.dart';
 
 export 'comment_endpoint.dart';
 export 'community_endpoint.dart';
@@ -56,5 +58,47 @@ class V1 with HttpHelper {
     });
 
     return Search.fromJson(res);
+  }
+
+  /// POST /admin/add
+  /// https://dev.lemmy.ml/docs/contributing_websocket_http_api.html#add-admin
+  Future<List<UserView>> addAdmin({
+    @required int userId,
+    @required bool added,
+    @required String auth,
+  }) async {
+    assert(userId != null);
+    assert(added != null);
+    assert(auth != null);
+
+    var res = await post('/admin/add', {
+      'user_id': userId,
+      'added': added,
+      'auth': auth,
+    });
+
+    List<dynamic> admins = res['admins'];
+    return admins.map((e) => UserView.fromJson(e)).toList();
+  }
+
+  /// GET /modlog
+  /// https://dev.lemmy.ml/docs/contributing_websocket_http_api.html#get-modlog
+  Future<Modlog> getModlog({
+    int modUserId,
+    int communityId,
+    int page,
+    int limit,
+  }) async {
+    assert(limit == null || limit >= 0);
+    assert(page == null || page > 0);
+
+    var res = await get('/modlog', {
+      if (modUserId != null) 'mod_user_id': modUserId.toString(),
+      if (communityId != null) 'community_id': communityId.toString(),
+      if (page != null) 'page': page.toString(),
+      if (limit != null) 'limit': limit.toString(),
+    });
+
+    return Modlog.fromJson(res);
   }
 }