+++ example/main.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:lemmy_api_client/lemmy_api_client.dart';
 
 void main() async {
@@ -6,9 +8,14 @@ void main() async {
   var lemmy = LemmyApi('dev.lemmy.ml').v1;
 
   // call methods that are named after op codes from the lemmy docs
-  var token = await lemmy.login(usernameOrEmail: 'asd', password: 'ads');
-  var messages =
-      await lemmy.getPrivateMessages(unreadOnly: true, auth: token.raw);
+  for (var i = 0; i < 200; i++) {
+    await lemmy.getPosts(
+        type: PostListingType.all, sort: SortType.active, communityId: 14799);
+    stdout.write("$i\r");
+  }
+  // var token = await lemmy.login(usernameOrEmail: 'asd', password: 'ads');
+  // var messages =
+  //     await lemmy.getPrivateMessages(unreadOnly: true, auth: token.raw);
 
-  print(messages);
+  // print(messages);
 }